move SecurityScheme model definition
