+++ packages/effect-http/src/Route.ts
@@ -14,6 +14,7 @@ import type * as Api from "./Api.js"
 import * as internal from "./internal/route.js"
 import type * as utils from "./internal/utils.js"
 import type * as RouterBuilder from "./RouterBuilder.js"
+import type * as SecurityScheme from "./SecurityScheme.js"
 import type * as ServerError from "./ServerError.js"
 
 /**
@@ -98,7 +99,7 @@ export type EndpointSchemasTo<E extends Api.Endpoint["schemas"]> = Types.Simplif
 /** @ignore */
 export type EndpointSecurityTo<Security extends Api.Endpoint["security"]> = Types.Simplify<
   {
-    [K in keyof Security]: Security[K] extends infer SS extends Api.SecurityScheme ? {
+    [K in keyof Security]: Security[K] extends infer SS extends SecurityScheme.SecurityScheme ? {
         token: Either.Either<ParseResult.ParseError, Schema.Schema.To<SS["decodeSchema"]>>
         securityScheme: SS
       } :
